+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 261:
 
 * Deduplicate `websocket::read_size_hint` definition
+* Fix UB in websocket read tests
 
 --------------------------------------------------------------------------------